Add auth-gen-token server option
We use auth-nocache option so every hour a renegotiation can cause custom configurations to fail because of a necessary input of the credentials. auth-gen-token option prevents this.

From the OpenVPN manual:
"The purpose of this is to enable two factor authentication methods, such as HOTP or TOTP, to be used without needing to retrieve a new OTP code each time the connection is renegotiated. Another use case is to cache authentication data on the client without needing to have the users password cached in memory during the life time of the session."